+++ crudapi/src/main/java/com/optimizertruck/crudapi/repository/ChauffeurRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChauffeurRepository extends JpaRepository<Chauffeur, Integer> {
+public interface ChauffeurRepository extends JpaRepository<Chauffeur, Long> {
 }
