+++ crudapi/src/main/java/com/optimizertruck/crudapi/repository/ContratRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContratRepository extends JpaRepository<Contrat, Integer> {
+public interface ContratRepository extends JpaRepository<Contrat, Long> {
 }
