+++ crudapi/src/main/java/com/optimizertruck/crudapi/repository/LivrerRepository.java
@@ -3,5 +3,5 @@ package com.optimizertruck.crudapi.repository;
 import com.optimizertruck.crudapi.model.Livraison;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LivrerRepository extends JpaRepository<Livraison, Integer> {
+public interface LivrerRepository extends JpaRepository<Livraison, Long> {
 }
