+++ crudapi/src/main/java/com/optimizertruck/crudapi/repository/LogisticienRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface LogisticienRepository extends JpaRepository<Logisticien, Integer> {
+public interface LogisticienRepository extends JpaRepository<Logisticien, Long> {
 
 
 /*    @Query("select l from Logisticien l where l.id = ?1")
