+++ crudapi/src/main/java/com/optimizertruck/crudapi/repository/MissionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface MissionRepository extends JpaRepository<Mission,Integer> {
+public interface MissionRepository extends JpaRepository<Mission,Long> {
 }
